+++ src/eslint/rules/index.ts
@@ -1,0 +1,13 @@
+/**
+ * @module @nhtio/adk/eslint/rules
+ *
+ * Barrel of the individual `@nhtio/adk` ESLint rule objects. Most consumers want the assembled
+ * plugin from `@nhtio/adk/eslint` instead; import individual rules from here only when composing a
+ * bespoke plugin or testing a single rule.
+ */
+
+export { default as requireValidatorAnyRequired } from './require_validator_any_required'
+export { default as noModelInToolHandler } from './no_model_in_tool_handler'
+export { default as thoughtPayloadRequiresReplayTag } from './thought_payload_requires_replay_tag'
+export { default as tokenEncodingRequiresContextWindow } from './token_encoding_requires_context_window'
+export { default as artifactToolForbidsArtifactConstructor } from './artifact_tool_forbids_artifact_constructor'